Fix bash range in get_repeated_scroll_cmd function
In #28, I changed the for loop in the `get_repeated_scroll_cmd` function
to use a bash built-in because `seq` is not portable (for example, it
doesn't exist on Solaris).  However, I must have been accidentally
testing under ZSH, because the range operator (`{1..x}`) does not
support variable interpolation under bash.  This commit changes the for
loop to using a math operator, which I have confirmed is compatible with
bash (at least back to version 3, which is quite old); and as a
built-in, is portable.